Organised files into lessons and activities

+++ lessons/chapter-6/intro/4-structarrays.c
@@ -9,7 +9,7 @@
 
 struct car
 {
-	int tank_capacity;
+	float tank_capacity;
 	int seat_capacity;
 	float city_mileage;
 	float price;
@@ -33,7 +33,7 @@ int main(void)
 	for (i = 0; i < 2; i++)
 	{
 		printf("Fuel capacity - car %d: ", i+1);
-		scanf("%d", &vehicle[i].tank_capacity);
+		scanf("%f", &vehicle[i].tank_capacity);
 
 		printf("Seat capacity - car %d: ", i+1);
 		scanf("%d", &vehicle[i].seat_capacity);
@@ -45,9 +45,10 @@ int main(void)
 		scanf("%f", &vehicle[i].price);
 	}
 
+	/** print information in the array */
 	for (i = 0; i < 2; i++)
 	{
-		printf("\nFuel capacity: \t%d\n", vehicle[i].tank_capacity);
+		printf("\nFuel capacity: \t%f\n", vehicle[i].tank_capacity);
 		printf("Seat capacity: \t%d\n", vehicle[i].seat_capacity);
 		printf("Auto mileage: \t%f\n", vehicle[i].city_mileage);
 		printf("Price: \t\t%3.2f\n", vehicle[i].price);
